Move tests inside source tree and update
